tooling/claat_export_images_py/requirements.txt
google-api-python-client==2.90.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
Pillow==9.5.0
requests==2.31.0
